feat: categories functions moved to external file
