+++ .eslintrc.json
@@ -11,6 +11,9 @@
         "ecmaVersion": "latest"
     },
     "rules": {
-        "linebreak-style": ["error", "windows"]  // changes the file to CRLF
+        "linebreak-style": ["error", "windows"], 
+        "import/extensions": [0, { "<js>": "always" }],
+        "no-plusplus": ["error", { "allowForLoopAfterthoughts": true }]
+
     }
 }
